WIP - test fixes
