+++ e2e/cypress/e2e/rendering.cy.ts
@@ -1,7 +1,7 @@
 describe("rendering", () => {
   beforeEach(() => {
     const port = Cypress.env("port")
-    cy.visit(`http://localhost:${port}/counter`)
+    cy.visit(`http://localhost:${port}/counter`).wait(1000)
   })
   it("displays the correct text in the site heading", () => {
     cy.get("header h1").should("have.text", "Hello World")
