+++ e2e/src/pages/document.tsx
@@ -1,3 +1,5 @@
+import "../style.css"
+
 export default function Document({ children }: { children?: JSX.Element }) {
   return (
     <html lang="en">
