+++ sandbox/ssg/src/pages/index.tsx
@@ -1,12 +1,11 @@
-import { signal } from "kiru"
-
-const text = signal("Hello World! 123")
+import { useState } from "kiru"
 
 export default function HomePage() {
+  const [count, setCount] = useState(2)
   return (
     <div>
-      <input bind:value={text} />
-      <h1>Welcome Home!</h1>
+      <button onclick={() => setCount(count + 1)}>Increment</button>
+      <span>Count: {count}</span>
       <p>This page is wrapped by the root layout only.</p>
       <p>You can see the navigation header and footer from the root layout.</p>
     </div>
